% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_var}
\alias{adjust_var}
\title{Adjust variance for sample size}
\usage{
adjust_var(n, sigma2_d)
}
\arguments{
\item{n}{vector of sample sizes}

\item{sigma2_d}{variance of the effect size}
}
\value{
vector of adjusted variances
}
\description{
Adjust variance for sample size
}
