% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_arg_len}
\alias{check_arg_len}
\title{Check arguments lengths}
\usage{
check_arg_len(len, ...)
}
\arguments{
\item{len}{integer vector of expected lengths}

\item{...}{list of arguments}
}
\description{
Check arguments lengths
}
