% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d_from_np}
\alias{d_from_np}
\title{Calculate p-value from sample size and effect size}
\usage{
d_from_np(n, lower.tail = FALSE, alpha = 0.05)
}
\arguments{
\item{n}{vector of sample sizes}

\item{lower.tail}{logical vector of 1L, FALSE indicates positive effect sizes and TRUE indicates negative effect sizes}

\item{alpha}{type I error rate}
}
\value{
vector of d-values
}
\description{
Calculate p-value from sample size and effect size
}
\details{
This calculation assumes a two-sided two-sample t-test with equal variance
}
