% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_methods.R
\name{discrete_weight}
\alias{discrete_weight}
\title{Discrete Weighting Function for Publication Bias}
\usage{
discrete_weight(p, w_pbs, alpha, detailed = FALSE)
}
\arguments{
\item{p}{Numeric vector of p-values}

\item{w_pbs}{Numeric scalar. Weight for probability of selection given P(p | p > alpha)}

\item{alpha}{Numeric scalar. Type I error rate}

\item{detailed}{Logical scalar. Return detailed output (default: FALSE)}
}
\value{
If detailed = FALSE, returns a vector of weights. If detailed = TRUE, returns a list with p, w, w_pbs, and alpha.
}
\description{
Discrete Weighting Function for Publication Bias
}
