% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{estimate_nb}
\alias{estimate_nb}
\title{Estimate the parameters of a alternative parametrized negative binomial distribution}
\usage{
estimate_nb(x, eps = 1e-04)
}
\arguments{
\item{x}{a numeric vector}

\item{eps}{a numeric scalar that determines the tolerance for convergence}
}
\value{
a named numeric vector with the estimated parameters (phi, mu)
}
\description{
Uses the method of moments to estimate the parameters of a negative binomial distribution
}
