% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{estimate_norm}
\alias{estimate_norm}
\title{Estimate the parameters of a normal distribution}
\usage{
estimate_norm(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a named numeric vector with the estimated parameters (mu, sigma)
}
\description{
Return the maximum likelihood estimates of the parameters of a normal distribution
}
\examples{
# Not run
x <- rnorm(100, mean = 0, sd = 0.5)
estimate_norm(x)
}
