% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{find_kde_bounds}
\alias{find_kde_bounds}
\title{Grid Bounds for 2-dimensional KDE}
\usage{
find_kde_bounds(
  x,
  bounds = c(0.005, 0.995),
  start_norm = c(0, 1),
  start_nb = c(10, 100)
)
}
\arguments{
\item{x}{data frame or matrix with colnames d and n}

\item{bounds}{vector of length 2 with lower and upper quantiles}

\item{start_norm}{vector of length 2 with starting values for ML-estimation for mu and sigma of normal distribution}

\item{start_nb}{vector of length 2 with starting values for ML-estimation for phi and mu of negative binomial distribution}
}
\value{
named vector of length 4 with lower and upper bounds for d and n
}
\description{
Computes the lower and upper bounds for the grid of the 2-dimensional KDE based on the absolute
minimum and maximum of the data and the quantiles of the fitted distributions.
}
\examples{
x <- cbind(d = rnorm(100, mean = 0.2, sd = 0.4), n = rnbinom(100, size = 10, mu = 100))
find_kde_bounds(x = x, bounds = c(0.005, 0.995), start_norm = c(0, 1), start_nb = c(10, 100))
}
\seealso{
\code{\link[=compute_loss]{compute_loss()}}
}
