% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{find_kde_limits}
\alias{find_kde_limits}
\title{Estimate the limits of the kernel density estimate based}
\usage{
find_kde_limits(x, pr = c(0.005, 0.995))
}
\arguments{
\item{x}{a numeric matrix with two columns (n, d)}

\item{pr}{A numeric vector of length 2 with the prs to estimate the limits}
}
\value{
a named list with the limits of the kernel density estimate
}
\description{
Estimate the bounds of the kernel density estimate based on the absolute minimum / maximum empirical data of the range of the data and the prs of the estimated normal and negative binomial distributions
}
