% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_2n}
\alias{get_2n}
\title{Split vector of total sample sizes into a matrix samples sizes by group}
\usage{
get_2n(n)
}
\arguments{
\item{n}{vector of total sample sizes}
}
\value{
matrix with two columns and len n rows of sample sizes
}
\description{
Split vector of total sample sizes into a matrix samples sizes by group
}
