% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{kl_div}
\alias{kl_div}
\title{Low-level KL divergence computation}
\usage{
kl_div(P, Q)
}
\arguments{
\item{P}{empirical distribution}

\item{Q}{theoretical distribution}
}
\value{
numeric value of the KL-divergence
}
\description{
This function computes the KL-divergence for discrete distributions. It handles cases where due to numerical precision,
p -> 0 and log(0) -> -Inf and Q -> 0 and P/Q -> Inf. In these cases, the function replaces 0 with the smallest positive number
and Inf with the largest positive number.
}
