% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named_list}
\alias{named_list}
\title{Selfnaming list
Taken from lme4::namedList()}
\usage{
named_list(...)
}
\arguments{
\item{...}{comma-separated arguments}
}
\description{
Selfnaming list
Taken from lme4::namedList()
}
