% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_from_nd}
\alias{p_from_nd}
\title{Calculate p-value from sample size and effect size}
\usage{
p_from_nd(n, d)
}
\arguments{
\item{n}{vector of sample sizes}

\item{d}{vector of effect sizes}
}
\value{
vector of p-values
}
\description{
Calculate p-value from sample size and effect size
}
\details{
This calculation assumes a two-sided two-sample t-test with equal variance
}
