% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_methods.R
\name{pbs_select}
\alias{pbs_select}
\title{Selection Bias based on Publication Bias}
\usage{
pbs_select(p, w_pbs, alpha)
}
\arguments{
\item{p}{vector of p-values}

\item{w_pbs}{weights for probability of selection given p(select | p > alpha)}

\item{alpha}{type I error rate}
}
\value{
boolean vector of selection
}
\description{
Selection Bias based on Publication Bias
}
