% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_joint_density.R
\name{plot_joint_density}
\alias{plot_joint_density}
\title{Plot selection biased joint density of sample size and effect size}
\usage{
plot_joint_density(
  n = seq(10, 250, 1),
  d = seq(-2, 2, 0.01),
  delta,
  exp_delta,
  sigma2,
  n_size,
  n_mu,
  alpha = 0.05,
  beta = 0.2,
  p_pbs,
  p_ssa
)
}
\arguments{
\item{n}{Vector of sample sizes}

\item{d}{Vector of effect sizes}

\item{delta}{Effect size mean}

\item{exp_delta}{Expected effect size}

\item{sigma2}{Effect size variance}

\item{n_size}{Negative binomial size parameter}

\item{n_mu}{Negative binomial mean parameter}

\item{alpha}{Type I error rate}

\item{beta}{Type II error rate}

\item{p_pbs}{weights of probability  selection given p(select | p > alpha)}

\item{p_ssa}{weights of probability selection given p(select | n < n_req)}
}
\value{
ggplot2 object
}
\description{
Plot selection biased joint density of sample size and effect size
}
\examples{
plot_joint_density(
  delta = 0.4, exp_delta = 0.4, sigma2 = 0.3, n_size = 4,
  n_mu = 120, p_pbs = 0.5, p_ssa = 0.5
  )
}
