% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pow2_add1}
\alias{pow2_add1}
\title{Check if a number is a power of 2 plus 1}
\usage{
pow2_add1(x)
}
\arguments{
\item{x}{a integer vector}
}
\value{
a logical vector
}
\description{
Check if a number is a power of 2 plus 1
}
