% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seap-package.R
\docType{package}
\name{seap-package}
\alias{seap-package}
\title{seap: Simulation based Estimation and Adjustment of Publication Bias in Meta-Analysis}
\description{
Provides functions for the simulation based estimation of publication bias and sample size planning in meta-analysis.
}
\author{
\strong{Maintainer}: Jan Luca Schnatz \email{schnatz@psych.uni-frankfurt.de}

}
\keyword{internal}
