% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{set_boundaries}
\alias{set_boundaries}
\title{Boundary Constraints for Simulated Annealing Optimization}
\usage{
set_boundaries(
  phi_n = c(1, 50),
  mu_n = c(50, 500),
  mu_d = c(-2, 2),
  sigma2_d = c(0.1, 2),
  delta_hat = c(0.01, 3),
  w_pbs = c(0, 1)
)
}
\arguments{
\item{phi_n}{integer vector of length 2, specifying the lower and upper boundaries for the dispersion parameter of the negative binomial distribution}

\item{mu_n}{integer vector of length 2, specifying the lower and upper boundaries for the mean of the negative binomial distribution}

\item{mu_d}{numeric vector of length 2, specifying the lower and upper boundaries for the mean of the normal distribution}

\item{sigma2_d}{numeric vector of length 2, specifying the lower and upper boundaries for the standard deviation of the normal distribution}

\item{delta_hat}{numeric vector of length 2, specifying the lower and upper boundaries for the expected effect size for sample size planning}

\item{w_pbs}{numeric vector of length 2, specifying the lower and upper boundaries for the publiation bias weights}
}
\value{
A list with two elements, lower and upper, each containing a named list of the lower and upper boundaries for the optimization algorithm.
}
\description{
Construct a list of lower and upper boundaries for the optimization algorithm.
}
