% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{set_start}
\alias{set_start}
\title{Set Starting Values for Loss Function Optimization via Simulated Annealing}
\usage{
set_start(
  phi_n = NULL,
  mu_n = NULL,
  mu_d = NULL,
  sigma2_d = NULL,
  delta_hat = NULL,
  w_pbs = 0.5
)
}
\arguments{
\item{phi_n}{numeric scalar, specifying the starting value for the dispersion parameter of the negative binomial distribution}

\item{mu_n}{numeric scalar, specifying the starting value for the mean of the negative binomial distribution}

\item{mu_d}{numeric scalar, specifying the starting value for the mean of the normal distribution}

\item{sigma2_d}{numeric scalar, specifying the starting value for the standard deviation of the normal distribution}

\item{delta_hat}{numeric scalar, specifying the starting value for the expected effect size for sample size planning}

\item{w_pbs}{numeric scalar (between 0 and 1), specifying the starting value for the publiation bias weights}
}
\value{
A named list of starting values for the optimization algorithm.
}
\description{
Construct a named list of starting values for the optimization algorithm.
}
