% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_methods.R
\name{sigmoid_weight}
\alias{sigmoid_weight}
\title{Sigmoid Weighting Function for Sample Size Adjustment}
\usage{
sigmoid_weight(
  n,
  delta_hat,
  slope_ssp = 5,
  alpha = 0.05,
  beta = 0.2,
  detailed = FALSE
)
}
\arguments{
\item{n}{vector of sample sizes}

\item{delta_hat}{expected effect size}

\item{slope_ssp}{discrimination parameter (defaults to 5)}

\item{alpha}{type I error rate (default = 0.05)}

\item{beta}{type II error rate (default = 0.2)}

\item{detailed}{return detailed output (default: FALSE)}
}
\value{
If detailed = FALSE, returns a vector of weights. If detailed = TRUE, returns a list with n, w_ssa, n_req, delta_hat, alpha, and beta.
}
\description{
Sigmoid Weighting Function for Sample Size Adjustment
}
