% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simetabias-package.R
\docType{package}
\name{simetabias-package}
\alias{simetabias-package}
\alias{_PACKAGE}
\title{simetabias: Simulation based Estimation of Publication Bias under Consideration of Sample Size Planning in Meta-Analysis}
\description{
Provides functions for the simulation based estimation of publication bias and sample size planning in meta-analysis.
}
\author{
\strong{Maintainer}: Jan Luca Schnatz \email{schnatz@psych.uni-frankfurt.de}

}
\keyword{internal}
