% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{simetabias}
\alias{simetabias}
\title{Estimate and Correct Publication Bias in Meta-Analysis under Consideration of Sample Size Planning}
\usage{
simetabias(emp_data, simetabias_control = simetabias_control())
}
\arguments{
\item{emp_data}{A dataframe or matrix with two columns: effect size (d) and sample size (n)}

\item{simetabias_control}{Control parameters \code{\link{simetabias_control}} for details.}
}
\description{
Estimate and Correct Publication Bias in Meta-Analysis under Consideration of Sample Size Planning
}
