% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{simetabias_control}
\alias{simetabias_control}
\title{Control of Important Simetabias Parameters}
\usage{
simetabias_control(
  bw = c("silverman", "scott", "sheather-jones", "ucv", "bcv"),
  n_grid = c(2^7 + 1, 2^7 + 1),
  pr = c(0.005, 0.995),
  k_sim = 10000,
  bounds = set_boundaries(),
  start = set_start(),
  alpha = 0.05,
  beta = 0.2,
  slope_ssp = 4,
  only_pbs = FALSE,
  trace = TRUE,
  hyperparameters = set_hyperparameters()
)
}
\arguments{
\item{bw}{bandwidth selection method. One of "silverman" (\link[stats]{bw.nrd0}), "scott" (\link[stats]{bw.nrd}),
"sheather-jones" (\link[stats]{bw.SJ}), "ucv" (\link[stats]{bw.ucv}) or "bcv" (\link[stats]{bw.bcv}).}

\item{n_grid}{integer vector of length 2, specifying the number evenly spaced grid points along each axis for density estimation.}

\item{pr}{numeric vector of length 2, specifying the lower and upper quantile for the CDF}

\item{k_sim}{integer vector of length 1, specifying the number of simulated primary studies for each iteration of the optimization algorithm}

\item{bounds}{a list generated with \code{\link{set_boundaries}} specifying the lower and upper bounds for the optimization algorithm}

\item{start}{a named list generated with \code{\link{set_start}} specifying the starting values for the optimization algorithm}

\item{alpha}{fixed type I error rate}

\item{beta}{fixed type II error rate}

\item{slope_ssp}{integer vector of length 1, specifying the discrimination parameter for the weighting function of sample size planning}

\item{only_pbs}{Logical scalar. If TRUE, the optimization algorithm will only consider the publication bias selection parameter. Default is FALSE.}

\item{trace}{Logical scalar. If TRUE, interim results are stored. Necessary for the plot function. Default is TRUE.}

\item{hyperparameters}{a list generated with \code{\link{set_hyperparameters}} specifying the hyperparameters for the simulated annealing optimization algorithm}
}
\description{
Contruct control structures for the simetabias publication bias estimation framework
}
