% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_meta.R
\name{simulate_meta}
\alias{simulate_meta}
\title{Simulate Meta-Analytical Data with Selection Bias (Publication Bias and Sample Size Adjustment Bias)}
\usage{
simulate_meta(
  k,
  n_size,
  n_mu,
  d_mu,
  d_sigma2,
  d_delta_hat,
  w_pbs,
  discr = 5,
  beta = 0.2,
  alpha = 0.05
)
}
\arguments{
\item{k}{number of primary studies in the 'population'}

\item{n_size}{size parameter of negative binomial distribution}

\item{n_mu}{mean parameter of negative binomial sample size distribution}

\item{d_mu}{mean parameter of normal effect size distribution}

\item{d_sigma2}{variance parameter of normal effect size distribution}

\item{d_delta_hat}{expected true effect size}

\item{w_pbs}{weight for probability of selection given p(select | p > alpha)}

\item{discr}{discrimination parameter for sigmoid function}

\item{beta}{type II error rate}

\item{alpha}{type I error rate}
}
\value{
matrix of simulated data with columns named n and d and row size depends on selection mechanism
}
\description{
This function simulates primary study data for a meta-analysis, incorporating selection biases such as publication bias
and bias due to sample size adjustment. The simulation is based on a population of potential primary studies with assumed
marginal distributions of effect sizes (normal distribution) and sample sizes (negative-binomial distribution).

The selection bias is introduced through two mechanisms:
\itemize{
\item Publication bias is introduced by a selection mechanism that chooses studies with p-values above a specified alpha-level
with a probability denoted by \code{w_pbs}.
\item Sample size adjustment bias is introduced by a selection mechanism that chooses studies with sample sizes below a required
sample size, given estimates from power-analysis with a specified power-level and expected effect size. The probability
of selection is denoted by \code{w_ssa}.
}
}
