% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speec-package.R
\docType{package}
\name{speec-package}
\alias{speec-package}
\title{speec: Simulation-based Publication Bias Estimation and Effect Size Correction}
\description{
Simulation-based Estimation Method of Publication Bias and Effect Size Correction under Publication Bias.
}
\author{
\strong{Maintainer}: Jan Luca Schnatz \email{schnatz@psych.uni-frankfurt.de}

}
\keyword{internal}
