% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_methods.R
\name{ssa_select}
\alias{ssa_select}
\title{Selection Bias based on Sample Size Adjustment}
\usage{
ssa_select(n, d_delta_hat, discr = 0.5, beta = 0.2)
}
\arguments{
\item{n}{vector of sample sizes}

\item{d_delta_hat}{expected effect size}

\item{discr}{discrimination parameter}

\item{beta}{type II error rate}
}
\value{
boolean vector of selection
}
\description{
Selection Bias based on Sample Size Adjustment
}
\examples{
n <- c(50, 20, 130)
ssa_select(n, 0.2, 0.2)
}
