% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{bivariate_kde}
\alias{bivariate_kde}
\title{Estimate a bivariate kernel density estimate}
\usage{
bivariate_kde(
  x,
  bw = c("silverman", "scott", "sheather-jones", "ucv", "bcv"),
  n_grid
)
}
\arguments{
\item{x}{a numeric matrix with two columns (n, d)}

\item{bw}{a character string specifying the bandwidth selection method}

\item{n_grid}{a integer vector (2L) specifying the number of equally spaced gridpoints in each direction (should be a power of 2 + 1)}
}
\value{
a matrix with the estimated densities
}
\description{
Estimate a bivariate kernel density estimate
}
\examples{
# Not run
k <- 100
X <- cbind(d = rnorm(k, mean = 0, sd = 0.5), n = rnbinom(k, size = 5, mu = 100))
bivariate_kde(X, bw = "silverman", n_grid = c(2^8 + 1, 2^8 + 1))
}
