% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{estimate_nb}
\alias{estimate_nb}
\title{Estimate the parameters of a alternative parametrized negative binomial distribution}
\usage{
estimate_nb(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a named numeric vector with the estimated parameters (phi, mu)
}
\description{
Estimate the parameters of a alternative parametrized negative binomial distribution
}
