% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{set_hyperparameters}
\alias{set_hyperparameters}
\title{Control of Important Hyperparameters for the Differential Evolution Optimization}
\usage{
set_hyperparameters(
  VTR = -Inf,
  strategy = 1,
  NP = 150,
  itermax = 1000,
  CR = 0.1,
  F = 0.9,
  trace = TRUE,
  initialpop = NULL,
  storepopfrom = 1,
  storepopfreq = 1,
  p = 0.2,
  c = 0,
  bs = TRUE,
  reltol = sqrt(.Machine$double.eps),
  steptol = itermax
)
}
\description{
Contruct control structures for the differential evolution optimization algorithm
}
