% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_methods.R
\name{ssa_select}
\alias{ssa_select}
\title{Selection Bias based on Sample Size Adjustment}
\usage{
ssa_select(n, delta_hat, slope_ssp = 0.5, beta = 0.2)
}
\arguments{
\item{n}{vector of sample sizes}

\item{delta_hat}{expected effect size}

\item{slope_ssp}{discrimination parameter}

\item{beta}{type II error rate}
}
\value{
boolean vector of length n
}
\description{
Selection Bias based on Sample Size Adjustment
}
